backend/src/requirements.txt
appdirs==1.4.3
boto==2.46.1
boto3==1.4.4
botocore==1.5.43
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-rednoise==1.0.5
django-storages==1.5.2
django-toolbelt==0.0.1
djangorestframework==3.6.2
djangorestframework-jwt=1.11.0
docutils==0.13.1
gunicorn==19.7.1
jmespath==0.9.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
static3==0.7.0
